finished update profile details

+++ login.php
@@ -1,35 +1,38 @@
 <?php 
     require_once 'core/init.php';
-
-    if( input::exists() ) {
-        if(token::check(input::get('token'))){
-            $validate = new validation();
-            $validate->check($_POST,array(
-                'username' => array(
-                    'required' => true 
-                ),
-                'password' => array(
-                    'required' => true 
-                )
-            ));
-        if($validate->passed()){
-            $user = new user();
-            $remember = (input::get('remember') == 'on') ? true : false ;
-            $login = $user->login(input::get('username'),input::get('password'),$remember);
-
-            if ($login){
-                redirect::to('index.php');
-            }else{
-                echo 'login failed ';
+    if( session::exists(config::get('session/session_name'))){
+        redirect::to('index.php');
+    }else{
+            if( input::exists() ) {
+                if(token::check(input::get('token'))){
+                    $validate = new validation();
+                    $validate->check($_POST,array(
+                        'username' => array(
+                            'required' => true 
+                        ),
+                        'password' => array(
+                            'required' => true 
+                        )
+                    ));
+                if($validate->passed()){
+                    $user = new user();
+                    $remember = (input::get('remember') == 'on') ? true : false ;
+                    $login = $user->login(input::get('username'),input::get('password'),$remember);
+        
+                    if ($login){
+                        redirect::to('index.php');
+                    }else{
+                        echo 'login failed ';
+                    }
+        
+                }else{
+                    foreach($validate->errors() as $error ){
+                        echo $error . '<br>';
+                        }
+                    }
+                }
             }
-
-        }else{
-            foreach($validate->errors() as $error ){
-                echo $error . '<br>';
-            }
-        }
     }
-}
 ?>
 <form action="" method="post">
     <div class="field">
